updated styles in seperate js
