+++ src/component/FooterComponent/Copyright/copyrightStyle.js
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  copyright: {
+    fontStyle: 'italic',
+    fontSize: '.8rem'
+  }
+}));
+
+export default useStyles;
