+++ src/component/FooterComponent/Home/Home.jsx
@@ -1,13 +1,7 @@
 import IconButton from '@material-ui/core/IconButton';
-import { makeStyles } from '@material-ui/core/styles';
 import HomeIcon from '@material-ui/icons/Home';
 import SocialLinks from '../Links/Social';
-
-const useStyles = makeStyles((theme) => ({
-  menuButton: {
-    marginRight: theme.spacing(2)
-  }
-}));
+import useStyles from './homeStyle';
 
 const Home = () => {
   const classes = useStyles();
