+++ src/component/FooterComponent/Home/homeStyle.js
@@ -1,0 +1,8 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles((theme) => ({
+  menuButton: {
+    marginRight: theme.spacing(2)
+  }
+}));
+export default useStyles;
