+++ src/component/FooterComponent/Logo/logoStyle.js
@@ -1,0 +1,9 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  logo: {
+    width: '5vw',
+    height: '5vh'
+  }
+}));
+export default useStyles;
