+++ src/component/FooterComponent/SocialmediaIcons/socialmediaStyle.js
@@ -1,0 +1,10 @@
+import { makeStyles } from '@material-ui/core/styles';
+
+const useStyles = makeStyles(() => ({
+  socialIconStyle: {
+    display: 'flex',
+    justifyContent: 'center',
+    alignItems: 'center'
+  }
+}));
+export default useStyles;
